change oitq: fix bug:prompt types
change plugins：
admin：friendly out
common：add Command exec，fix recall bug

change services:
console:fix bug (vite dev mode)
task:fix bug (creator field)

+++ services/task/src/model/task.ts
@@ -2,7 +2,7 @@ import {DataTypes,Model} from '@oitq/service-database'
 export const Task={
     name:DataTypes.STRING,
     desc:DataTypes.TEXT,
-    creator:DataTypes.INTEGER
+    creator:DataTypes.BIGINT
 }
 export interface Task{
     id:number
